requirements.txt
streamlit==1.39.0
litellm==1.51.0
python-dotenv==1.0.1
requests==2.32.3
fastapi==0.115.2
uvicorn==0.31.1
pytest==8.3.3
